requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
electricitycostcalculator==0.1.0
googleapis-common-protos==1.5.9
grpcio==1.22.0
grpcio-tools==1.16.1
holidays==0.9.10
hupper==1.6.1
idna==2.8
kiwisolver==1.1.0
lxml==4.3.4
matplotlib==3.1.1
numpy==1.16.2
pandas==0.24.2
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.7.1
pymortar==1.0.4
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rfc3339==6.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
translationstring==1.3
urllib3==1.25.3
venusian==1.2.0
WebOb==1.8.5
xbos-services-getter==0.0.2.dev44
zope.deprecation==4.4.0
zope.interface==4.6.0
